\name{VAN-package}
\alias{VAN-package}
\alias{VAN}
\docType{package}
\title{ Differential variability analysis in networks }

\description{
This package enables the identification of hubs whose 
association with their partners (interactors) changes 
across biological conditions.
}
 
\details{
\tabular{ll}{
Package: \tab VAN\cr
Type: \tab Package\cr
Version: \tab 1.0.0\cr
Date: \tab 2012-12-01\cr
License: \tab LGPL(>=2.1)\cr
LazyLoad: \tab yes\cr
}

}
\author{
Vivek Jayaswal, Sarah-Jane Schramm and Yee Hwa (Jean) Yang
Maintainer: Vivek Jayaswal <vivek@maths.usyd.edu.au>
}

\keyword{ package }
